==> step_1_data_cleaning.ipynb <==
import csv
import pandas as pd
import re
import sqlite3
from sqlite3 import Error
import ast 
import math
import os
import json
from datetime import datetime

receipts_df = pd.read_json('receipts.json', lines=True)
brands_df = pd.read_json('brands.json', lines=True)
users_df = pd.read_json('users.json', lines=True)

### Clean Receipts
- 1): Remove dictionary key from ids
- 2): Change date format
- 3): Make sure numerical columns have type as numeric
- 4): use encoding = utf-8 for next step schema design

def clean_receipts_dataset(df):
    
    df['_id'] = df['_id'].apply(lambda x: x['$oid'] if isinstance(x, dict) and '$oid' in x else x)
    df['userId'] = df['userId'].apply(lambda x: x['$oid'] if isinstance(x, dict) and '$oid' in x else x)
    
    # Clean date columns
    date_columns = ['createDate', 'dateScanned', 'finishedDate', 'modifyDate', 'pointsAwardedDate', 'purchaseDate']
    for col in date_columns:
        if col in df.columns:
            df[col] = df[col].apply(lambda x: datetime.fromtimestamp(x['$date'] / 1000) if isinstance(x, dict) and '$date' in x else x)
    
    # Clean numeric columns
    numeric_columns = ['bonusPointsEarned', 'pointsEarned', 'purchasedItemCount', 'totalSpent']
    for col in numeric_columns:
        if col in df.columns:
            df[col] = pd.to_numeric(df[col], errors='coerce')
        
    df.to_csv('cleaned_receipts.csv', index=False, encoding='utf-8')
    return df

cleaned_receipts_df = clean_receipts_dataset(receipts_df)
cleaned_receipts_df.head()

### Clean Users
- 1): Remove dictionary key from ids
- 2): Change date format
- 3): Change active column from T/F to 1/0
- 4): Remove duplicated rows from User table
- 5): use encoding = utf-8 for next step schema design

def clean_users_dataset(df):
    
    df['_id'] = df['_id'].apply(lambda x: x['$oid'] if isinstance(x, dict) and '$oid' in x else x)

    # Clean date columns
    date_columns = ['createdDate', 'lastLogin']
    for col in date_columns:
        if col in df.columns:
            df[col] = df[col].apply(lambda x: datetime.fromtimestamp(x['$date'] / 1000) if isinstance(x, dict) and '$date' in x else x)
    
    df['active'] = df['active'].astype(int)
    df = df.drop_duplicates(subset=['_id'])
    df.to_csv('cleaned_users.csv', index=False, encoding='utf-8')
    return df

cleaned_users_df = clean_users_dataset(users_df)
cleaned_users_df.head()

### Clean Brands
- 1): Remove dictionary key from ids
- 2): use encoding = utf-8 for next step schema design

def clean_brands_dataset(df):
    
    df['_id'] = df['_id'].apply(lambda x: x['$oid'] if isinstance(x, dict) and '$oid' in x else x)
    
    def extract_oid(value):
        if isinstance(value, dict) and '$id' in value and isinstance(value['$id'], dict) and '$oid' in value['$id']:
            return value['$id']['$oid']
        return None

    df['cpg'] = df['cpg'].apply(extract_oid)
    
    df.to_csv('cleaned_brands.csv', index=False, encoding='utf-8')
    return df

cleaned_brands_df = clean_brands_dataset(brands_df)
cleaned_brands_df.head()
