==> step_4_data_quality_issue.ipynb <==
import csv
import pandas as pd
import re
import sqlite3
from sqlite3 import Error
import ast 
import math
import os
import numpy as np

#Original datasets -df
receipts_df = pd.read_json('receipts.json', lines=True)
brands_df = pd.read_json('brands.json', lines=True)
users_df = pd.read_json('users.json', lines=True)

#Cleaned datasets -cdf
receipts_cdf = pd.read_csv("cleaned_receipts.csv")
users_cdf = pd.read_csv("cleaned_users.csv")
brands_cdf = pd.read_csv("cleaned_brands.csv")

### Receipts

receipts_df.head()

#### 1): Check how many null values are in each column

Noticed that for **"rewardsReciptItemList"** column, which were used to store items information, **had a lot of null values (440)**; this might be an issue when analyze the item counts/price/information etc.

receipts_df.isna().sum()

We can see that 434 were related to status of "submitted", **and 6 were related to others (which I guess they suppose to have an item list but do not)**

receipts_df[receipts_df['rewardsReceiptItemList'].isnull()].groupby('rewardsReceiptStatus').size()

conn = sqlite3.connect("receipts.db")
cursor = conn.cursor()

#### 2): Check missing information in rewardsReceiptItemList column

If we take a closer look at item within item list, there are a few issue raised when creating the relational database schema:
- 1): item has different keys combination within it's dictionary, **not all item has a barcode (missing unique itentifier for some items)** thus some item will not be found in the item table if it doesn't have a barcode
- 2): **not all item has a brand code, 148 items missing brand code even after I use "description" to map some of the brand code to the item table**, moreover, since brand code is not the unique itentifier in brand table (brand id is), I will need to use brand code to find brand id in order to connect these 2 tables, but brand code might corresponding to different brand ids
    - this might casue issues like lower item counts by brand code

cursor.execute("""
select count(*) 
from item
where brand_code is null
""")
answer = cursor.fetchall()
answer

### Brands

brands_df.head()

If we take a closer look into brand table, **234 brandCode were missing**, **35 were stored nothing/space as brand so will be cleaned as null value in cleaning step 1, total were 269 missing**. This was related to the item issue mentioned previously

brands_df['brandCode'].value_counts()

cursor.execute(f"""
select count(brand_id)
from brand
where brand_code is null
""")
answer = cursor.fetchall()
answer

brand id is unique identifier, same brand name might has multiple brand ids, same barcode might has multiple brand ids, this might be very confusing when join to another table to find brand

cursor.execute(f"""
select count(distinct(brand_id)), count(distinct(barcode)), count(distinct(name)), count(distinct(brand_code))
from brand
""")
answer = cursor.fetchall()
answer

### Users

users_df.head()

Over 200 duplicated rows in users table, cleaned in step 1 data notebook

users_df.count()

users_cdf.count()

**around 148 receipt** doesn't have a corresponding user in user table

cursor.execute(f"""
select count(receipt_id)
from receipt r
left join user u
on r.user_id = u.user_id
where u.user_id is null
""")
answer = cursor.fetchall()
answer
